career-digital-twin/requirements.txt
# Linting and Code Quality
flake8==6.0.0
black==23.3.0
isort==5.12.0
mypy==1.3.0
pylint==2.17.0
pytest==7.3.1
python-dotenv==1.0.0

# Main dependencies
flask==2.3.3
flask-cors==4.0.0
openai==0.28.0